refactor: migrate deck swiper page
