+++ src/features/card/index.ts
@@ -1,3 +1,6 @@
+export { BackText } from "./components/BackText";
+export { CardOverlay } from "./components/CardOverlay";
+export { FrontText } from "./components/FrontText";
 export { CardFormContainer } from "./containers/CardFormContainer";
 export { CardListContainer } from "./containers/CardListContainer";
 export { CardViewContainer } from "./containers/CardViewContainer";
