+++ src/features/study/index.ts
@@ -1,10 +1,14 @@
 export { DeckStartContainer } from "./containers/DeckStartContainer";
-export { DeckSwiperContainer } from "./containers/DeckSwiperContainer";
+export { Controller, type ControllerProps } from "./components/Controller";
+export { SwipeButtonList, type SwipeButtonListProps } from "./components/SwipeButtonList";
 export {
   discardStudySessionsMissingDecks,
   removeStudySession,
   touchStudySession,
 } from "./commands/studySessionCommands";
 export { useStudyHydrated } from "./hooks/useStudyHydrated";
+export { useStudyActions } from "./hooks/useStudyActions";
+export { useStudyControllerState } from "./hooks/useStudyControllerState";
 export { useStudySessions } from "./hooks/useStudySessions";
-export type { StudySession } from "./state/studyStore";
+export { useStudyStore } from "./hooks/useStudyStore";
+export { selectStudySessionForRoute, studyStore, type StudySession } from "./state/studyStore";
